TheWardrobe.API.ClothesClassification/requirements.txt
requests
Pillow
Flask
tensorflow
tensorflow_hub